firmware: use "Usage Multi Mode Indicator" selector for led
this allows the client to request the led to blink or flash without
having to control that manually.

golang library not supporting this yet.

this also improves the led handling and the firmware code in general.

when turning on the device, the led will stay on until the host sets its
address.