firmware: add automatically generated serial number
the serial number is generated from non-initialized flash memory, which
is not very reliable, and stored in eeprom. we may change this algorithm
at some point.

also, the code that implements this won't fit the attiny2313. devices
using it won't have serial numbers, and are limited to one unit
connected per computer.