firmware: display: add a usb out report to clear display with delay.
the delay is set in milliseconds, limited to 65535ms. this is the maximum
delay applied when clearing the display: if any write operation happens while
the delay is ongoing, the display is erased immediately before writing the
new data. no new data is delayed or lost, making this feature very suitable for
notifications.